refactor: audit 관련 클래스를 `global.audit` 패키지로 이동
- audit와 관련된 모든 클래스를 `global.audit` 패키지로 이동했습니다.
- 새로운 패키지 구조를 반영하도록 import 문과 참조를 업데이트했습니다.
