+++ src/main/java/com/wootecam/festivals/global/audit/AuditingConfig.java
@@ -1,9 +1,9 @@
-package com.wootecam.festivals.global;
+package com.wootecam.festivals.global.audit;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @Configuration
 @EnableJpaAuditing
-public class JpaConfig {
+public class AuditingConfig {
 }
